Moved styles to Feed component
