+++ client/components/Feed.jsx
@@ -1,16 +1,50 @@
-import React, { useEffect, useState } from 'react';
-// import ReactDOM from 'react-dom';
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Button from '@mui/material/Button';
+import CameraIcon from '@mui/icons-material/PhotoCamera';
+import Card from '@mui/material/Card';
+import CardActions from '@mui/material/CardActions';
+import CardContent from '@mui/material/CardContent';
+import CardMedia from '@mui/material/CardMedia';
+import CssBaseline from '@mui/material/CssBaseline';
+import Divider from '@mui/material/Divider';
+
+import Drawer from '@mui/material/Drawer';
+import Grid from '@mui/material/Grid';
+import List from '@mui/material/List';
+import ListItem from '@mui/material/ListItem';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import InboxIcon from '@mui/icons-material/MoveToInbox';
+import MailIcon from '@mui/icons-material/Mail';
+import Stack from '@mui/material/Stack';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import Link from '@mui/material/Link';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
 import FeedCodeBlock from './FeedCodeBlock.jsx';
 
-export default function Feed(props) {
-  // React hooks for state - store the data from the database
-  const [codeBlocks, setCodeBlocks] = useState([{       topic: 'javascript',       date: '8/10/2021',       upvotes: 5,       downvotes: 20,       title: 'This is a sample postMessage',       issue: 'This is the issue',       tried: 'this is what I tried',        cause: 'I am not sure',       code: '<code>',       replies: [{content:'This is reply 1', code: 'sample reply 1 code', upvotes:10 , downvotes: 5, date: '8/15/2020',}, {content: 'This is reply 2', code: 'this is sample 2 reply code', upvotes: 1, downvotes: 2, date: '8/20/2021',}]       }]);
-  const [post, setPost] = useState();
-  // const [posts, setPosts] = useState([]); //this was taken from CreatePost - same purpose as codeBlocks/setCodeBlocks
 
+const theme = createTheme();
+
+const headingFont = createTheme({
+  typography: {
+    fontFamily: [
+      'PT Mono',
+      'monospace'
+    ].join(','),
+    fontSize: 24,
+  }
+})
+
+
+export default function Feed(props) {
+
+  const [codeBlocks, setCodeBlocks] = React.useState([]);
 
   // update state that we fetch
-
   function getFeed() {
     fetch(`/api/gettopic/${props.topic}`) //removed /topic
       .then((res) => res.json())
@@ -20,26 +54,60 @@ export default function Feed(props) {
       .catch((err) => console.log(err));
   }
 
-  useEffect(() => {
+  React.useEffect(() => {
     getFeed();
   },[props.topic]);
-  
+
   // create codeblock components and save them in an array called 'codeBlockEl'
   const codeBlockEl = codeBlocks.map(post => {
     return (
-      <FeedCodeBlock 
-        key={post._id} 
-        post={post}
-        clickHandler = {props.clickHandler}
-      />
+      <Grid item key={post} xs={12} sm={6} md={4}>
+        <FeedCodeBlock sx={{ height: '100%', display: 'flex', flexDirection: 'column'}}
+          key={post._id} 
+          post={post}
+          clickHandler = {props.clickHandler}
+        />
+      </Grid>
     );
   });
-
-  // returns code block cards
   return (
-    <div id='feed'>
-      <h1>Welcome to Rate-My-Code</h1>
-      {codeBlockEl}
-    </div>
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+
+      <main>
+        {/*Hero*/}
+        <Box
+          sx={{
+            bgcolor: 'Background.paper',
+            pt: 8,
+            pb: 6,
+          }}
+        >
+          <Container maxWidth='sm'>
+            <ThemeProvider theme={headingFont}>
+              <Typography
+                component='h1'
+                variant='h2'
+                align='center'
+                color='text.primary'
+                gutterBottom
+              >
+              Welcome to RATE-MY-C0D3
+              </Typography>
+            </ThemeProvider>
+            <Typography variant='h5' align='center' color='text.secondary' paragraph>
+              Rate-My-Code is a supportive online community where you can post blocks of code that are not acting as you expect them to.
+            </Typography>
+
+          </Container>
+        </Box>
+        <Container sx={{ py: 8 }} maxWidth="md">
+          {/* End hero unit */}
+          <Grid container spacing={4}>
+            {codeBlockEl}
+          </Grid>
+        </Container>
+      </main>
+    </ThemeProvider>
   );
 }
