+++ client/containers/MainContainer.jsx
@@ -18,7 +18,7 @@ import InboxIcon from '@mui/icons-material/MoveToInbox';
 import MailIcon from '@mui/icons-material/Mail';
 
 import classes from './MainContainer.module.css';
-import ThemedFeed from '../components/ThemedFeed.jsx';
+import Feed from '../components/Feed.jsx';
 import CreatePost from '../components/CreatePost.jsx';
 import PostView from '../components/PostView.jsx';
 import './custom.scss';
@@ -67,10 +67,27 @@ export default function MainContainer() {
             <Divider />
           </Drawer>
         </div>
+        {/* This makes the createPost button */}
+        <div>
+          <Link to="/home/createpost">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="32"
+              height="32"
+              fill="currentColor"
+              className="bi bi-plus-circle"
+              viewBox="0 0 16 16"
+            >
+              <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
+              <path d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z" />
+            </svg>
+          </Link>
+        </div>
+
         <div className={classes.righttwothirds}>
           <Switch>
             <Route path="/home" exact>
-              <ThemedFeed clickHandler={handlePostClick} topic={topic} />
+              <Feed clickHandler={handlePostClick} topic={topic} />
             </Route>
             <Route path="/home/createpost">
               <CreatePost />
@@ -82,22 +99,7 @@ export default function MainContainer() {
         </div>
       </div>
 
-      {/* This makes the createPost button */}
-      <div>
-        <Link to="/home/createpost">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="32"
-            height="32"
-            fill="currentColor"
-            className="bi bi-plus-circle"
-            viewBox="0 0 16 16"
-          >
-            <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
-            <path d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z" />
-          </svg>
-        </Link>
-      </div>
+
     </Container>
   );
 }
